pyproject.toml
[project]
name = "gpt-trade"
version = "0.1.0"
description = "Event-driven algorithmic trading system"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
dependencies = [
    "pydantic",
    "fastapi",
    "uvicorn",
    "httpx",
    "google-genai",
    "jsonschema",
    "pandas",
    "numpy",
    "sqlite-utils",
    "python-dotenv",
    "pyyaml",
    "alpaca-trade-api",
    "aiohttp",
    "structlog",
    "schedule",
    "pytest",
    "pytz>=2025.2",
    "urllib3>=1.26,<2",
]
requires-python = ">=3.11"
license = { text = "MIT" }

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["gpt_trade*"]
